batch: Split ownership helper modules
Ownership metadata had become the shared home for claim range helpers and
replacement-unit normalization. Remapping, translation, and unit rebuild
code had to import private names from the metadata model.

The new claim helper module owns source-line range parsing, formatting, and
presence claim construction. A replacement-unit module owns validation and
coalescing for persisted replacement metadata.

This commit moves those helpers behind public module boundaries and updates
callers to depend on those boundaries.

Ownership model code keeps the data classes, serialization paths, detached
buffer handling, and metadata acquisition logic.
